+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### 3.8.8 (11/20/2015 - )
 * uri: simplify URIBuilder to match our use cases and performance tuning
+* util: ByteBuf put byte[]
+* refactory: performance tuning for html/uri encode/decode
 
 ### 3.8.7 (11/20/2015)
 * bug: fixed QueryParam encoding to encode +/?/=